Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
+
 class QuestionsController extends Controller
 {
     /**
@@ -56,7 +57,8 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-        //
+        $question->increment('views');
+        return view('questions.show', compact('question'));
     }
 
     /**
@@ -67,7 +69,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-       return view('questions.edit',compact('question'));
+       if (\Gate::denies('update-question', $question)) {
+           abort(403, 'Access Denied');
+        }
+        return view('questions.edit', compact('question'));
     }
 
     /**
@@ -79,6 +84,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+           abort(403, 'Access Denied');
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('/questions')->with('success','Your question has been updated');
     }
@@ -91,6 +99,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+           abort(403, 'Access Denied');
+        }
        $question->delete();
        return redirect('/questions')->with('success', 'You have successfully deleted the question');
     }
